Fix processing pipe magic word in bailing wikilinks
The special case in processSpecialMagicWord is only meant to support
parsertests where the pipe magic word is encountered nested in templates
and needs to be emitted as a table data cell.  However, because bailing
on wikilinks wasn't setting the toplevel flag and the default for a
nested pipeline is to set toplevel as false, pipe magic words were
always being emitted as tds.

The fix switches the condition to use inTemplate instead.  atTopLevel
means something different than when the code in processSpecialMagicWord
was originally written and inTemplate better captures the intention.

The change to the knownFailures list is because, prior to the fix, an
empty td was being dropped when doing tokensToString to get the link
target.  However, now that pipe is properly being emitted, we bail on
the wikilink as unsupported from T226523 and T290526.  The case of the
pipe magic word should be easier to handle in grammar than the general
case of a template expanding with a pipe char since we don't have to
wait for the template expansion to know the contents and, indeed,
wikilink_content accepts the pipe rule, so fixing that in the short term
seems feasible.

Bug: T387237
Bug: T290526
Change-Id: I022f0ac400c23641009977cd47ea72ee8833da58