Support {{!}} as a top level wikilink seperator
As outlined in I022f0ac400c23641009977cd47ea72ee8833da58, the grammar
mostly already supported {{!}} as a separator in wikilinks.

The more general case of a template in wikilink target position
expanding to a pipe still isn't supported.

Parser tests were updated to show the newly passing case.

The regression test output was changed to reflect that bailing on the
pipe trick would need to emit a template token to roundtrip cleanly.

Bug: T290526
Change-Id: If30a3351d82a3f9dcef3505d1929a47a5a8848fd