Make whitespace normalization of extension options configurable
Extension options are currently normalized by the extension
handler: sequences of new lines, tabs and spaces are replaced by
a single whitespace, and the strings are trimmed.

Some extensions may want to rely on white space in the option
arguments: this is the case of Kartographer and the "text" attribute,
and some uses of Cite rely on whitespace to distinguish between
different citation names.

This patch proposes a way to give whitespace control to the
extensions, while keeping the current normalization behaviour as
a default. A companion patch on Cite shows how this might be
used.

Bug: T386713
Change-Id: I792ba7785433ba8d2d9ee2c8427f488c986cd3a1