Add ::processAttributeEmbeddedDom() methods
The {ParsoidExtensionAPI,ContentUtils}::processAttributeEmbeddedDom()
method provides a alternative to ::processAttributeEmbeddedHTML()
which accepts a callback taking a DocumentFragment, instead of an
HTML string.

This avoids serialization/deserialization overhead when the
embedded HTML is stored natively as a DocumentFragment (not a
string), which I4e8e56ddf8a103e3172b13bde62b4c6478d46e30 will
allow -- but it is also simply a more convenient interface in
most cases since the client doesn't need to manually parse the
HTML string; see Iecdaa41c1cacecfaea1c7c82fd9f71a16a7255b7 for
example.

Similarly, ExtensionTagHandler::processAttributeEmbeddedDom() is
added as an alternative to ::processAttributeEmbeddedHTML() when
the embedded contents are natively a DocumentFragment, not a
string.  Although I've converted Parsoid's internal extensions to
override this new method, other extensions can/should implement
whichever is most convenient for the format the embedded HTML is
stored in.

Change-Id: If14a86645b9feb5b3d9503e3037de403e588d65d