Improve debugabbility of JsonString decoding issues on data-mw
DataMw::newFromJsonArray applies array_map to its 'parts' item, which
can apparently be a string in some cases when coming back from
VisualEditor. In this patch, we re-throw on TypeError exception with
more context to improve debuggability.

The impact of this production warning is limited in scope (seen on a
single page in the past 10+ days), but it feels worth trying to
investigate a bit more before blindly adding a "is_array" to avoid
throwing.

Bug: T388160
Change-Id: I8b31768765fdb0785b638926f4a6af40db67b244