Rich attribute support, phase 1b
This patch extends the Rich Attribute support to included embedded
DocumentFragments, which are stored live in the NodeData (or in an
object stored in the NodeData, like DataMw). This patch corresponds to
"phase 1b" of the proposal at:

https://www.mediawiki.org/wiki/Parsoid/MediaWiki_DOM_spec/Rich_Attributes

As an experimental feature, this patch also introduces "fragment bank"
serialization, which avoids HTML-escaping HTML strings in order to
store them in attribute values; instead each HTML fragment is
serialized to a <template> element in the <head> of the document, and
only a reference to the ID of that <template> element is serialized in
the attribute.  In addition to providing a compact (and
human-readable) encoding by avoiding unnecessary entity-escapes, this
also keeps all embedded DocumentFragments as proper children of the
Document in the DOM, allowing certain queries and traversals to be
done without special knowledge of which attributes contain serialized
HTML.

The --fragmentbank option can be given to parse.php to use the
fragment bank serialization, but to see it in action first a patch
must be merged to actually use a DocumentFragment in the
NodeData/DataMw, like Id0ef549560eec2c1609e5826e20068579e64b5ee.

Bug: T348161
Change-Id: I4e8e56ddf8a103e3172b13bde62b4c6478d46e30