Throw on error in json_decode of knownFailures
If the knownFailures file is invalid in some way, its decoding silently
returns null, which is very confusing when an unrelated patch suddenly
starts emitting a lot of parserTests failures.
This patch adds the JSON_THROW_ON_ERROR to this path so that we make
sure that the knownFailures file is indeed loaded when it exists.

Change-Id: I10f72ad2454ba54013b3fc4e2c64eccb3c3a70bf