Don't use Utils::clone() in AttributeExpander; add token __clone methods
Utils::clone() is not supposed to be used on anything other than
"arrays or plain objects", however it was being used in AttributeExpander
on value arrays containing DataParsoid and DataMw objects, which led
to Bad Things happening when DataMw objects containing DocumentFragments
were cloned.  WikiLinkHandler was also using Utils::clone() on an
array of KV objects.

Add Utils::cloneArray() which ought to be safe for use on arrays with
any type of contents (provided that the contents support PHP's `clone`
operator and/or properly implement `__clone()`), and add implementations
of Token::__clone() (and subclasses), KV::__clone(), and
TempData::_clone().  Fixup DataParsoid::clone() to avoid using
Utils::clone().

Utils::clone() should probably be deprecated and replaced at all other
places where it is used, but this is a good start.

Bug: T390499
Change-Id: Ie1b6e4ff611292a61a892363c8010acb267cd525