FragmentHandler support: Don't pass a empty string if no args are given
{{#foo}} has zero arguments, while {{#foo:}} has one argument which is
an empty string.  Distinguish these two cases and ensure that we pass
zero arguments where that is appropriate.

This patch is a little bit hacky, since the proper solution will involve
T391063 (magic variable support), since for no-hash parser functions
the {{foo}} syntax corresponds to a "magic variable" invocation.
We'll clean this up further when we tweak the tokenizer for T390344
as well.

Change-Id: I92c929f89f8056d221334a26ed4edee942ddd913