StripState::addAllFrom() bug fix
PHPUtils::pushArray() only works on PHP lists (integer indexes), not
associative arrays.  What we want is the `+` operator; even
`array_merge` is not quite right because it would renumber numeric
indexes.  Even though we currently assign keys so that they should all
be non-numeric, play it safe and spell out the merge operation instead
of trying to be clever.

Change-Id: I475e70665f61033466a657290f2296806dc94d73