Make Tokens JsonCodecable
Replace a bunch of old manual Token serialization/deserialization
code with JsonCodec methods.  Keep things \JsonSerializable for
any old logging code which still does a manual json_encode(), but
using JsonCodec you can reliably deserialize the objects as well.

Deserialization is really only used in the TokenUtilsTest unit
test, but it's not hard to keep it working with a more modern
framework.

Change-Id: I4a4dacdb8631df9cf9a3ca04b6e02c40ccd3597a