ParsoidLogger: improve formatting of log messages
By supporting better formatting in the logger, we don't have to create
closures which do this formatting for us.  We use XMLSerializer for
Node types, and JsonCodec for all other non-string types.  Further,
our closures can return object or Node types now, which we can
format in the standard way.

Change-Id: Ic840d50c69f9f2d6ae6ddd36cccdd1a471a42895