Add Env::trace() to more efficiently short-circuit tracing when disabled
This is a mechanical transformation from
   ->env->log( "trace/...", ...)
to
   ->env->trace( "...", ...)

In order to avoid a regexp match in the common case (in production)
when no tracing is enabled.

The old style `->env->log( "trace/...", ...)` still works, it is just
a tiny bit slower to execute in the no-tracing case.

There are some additional cases where we call `->log` with a computed
log level that happens to be tracing; I will do those in a follow up
patch in order to keep this (large) patch strictly mechanical.

Change-Id: I1557ce1d359653dfb385795d01dd231376f9b3ea