TokenHandlerResult: Get rid of useless validation
Any errors in return values will crash future handlers and there
is no reason to incur this penalty for every handler call's
return value.

We have sufficient testing in place at this point to catch crashers
before they hit production.

The benefits are somewhat marginal, but followup patches should be
able to get rid of this class altogether and have handlers stick to
the following convention:
* null for unchanged results.
* token array for modified results.

Change-Id: I036f4a3a5d1b5ce33f855d531af371f389f4e23d