Turn dsr info logs into traces
Info logs don't make it to logstash and we aren't acting on the
information, it's creating noise in phpunit output.

However, it's still useful to capture the information when tracing.

If anything, given the importance of dsr info for roundtripping, we
should find a way to turn these into errors.

Bug: T358445
Change-Id: Icb6babd67aae49c5a6751719c4c3ce54cfc1d1f2