New PreprocTk tokens for "unexpanded preprocessed text"
In this patch we add `PreprocTk` tokens and associated tokenizer rules
while represent "unexpanded preprocessed text".  That is, the contents
of the `PreprocTk` tokens are `list<string|PreprocTk>` and *not*
fully-tokenized wikitext.  In order for the tokens to be parsed, they
must be expanded and retokenized.

However, this format is perfect for extension tags and transclusions
where it is desired to separate the arguments but the actual argument
handling is delegated elsewhere.  It preserves all of the original
text (including whitespace) so that parser functions and other
fragment handlers can access the arguments in whatever form they
wish.

In this patch the new `preproc_piece` and `preproc_pieces` rules which
create these tokens are not yet directly used by the tokenizer.
Future patches will change the tokenization of wikitext which starts
`{{#...` to use these rules (T394834).  Eventually we may use
`preproc_piece` to tokenize other/all curly-brace constructs
(T394836).

Note that this patch is consistent with the legacy preprocessor in
that heading indexes are assigned during preprocessing, regardless
of whether the headings actually made it into the final token stream.

The core preprocessor emits special "ignore" tokens for <noinclude>,
<includeonly>, and <onlyinclude> elements, with the exact text ignored
depending on whether we are inside a template or not.  The new
preproc_piece rules match the core preprocessor behavior with a
PreprocIgnoreTk token type.

In wikitext constructs like:
```
{{1x|
=== Heading ===
}}
```
The heading is not processed as a named argument, despite containing
an equal sign, because it is tokenized as a separate "possible heading"
token.  Match this behavior of the legacy parser in Parsoid with a
PreprocHeadingTk token type.

Note that the core preprocessor handles double-bracket `[[ ... ]]`
wikilinks as a token, but *not* single-bracket `[ ... ]` extlinks.
This behavior is revealed in how (eg) template boundaries are
determined.  This behavior is matched in Parsoid: only double-bracket
constructs become PreprocTk tokens.

Bug: T393391
Change-Id: I8c7b535d8424fa70ec4d813118d18479cfd12203