Hoist context check out of the loop
Context doesn't change but presumably the compiler knows this and isn't
really an efficiency gain.

Also, renames subpipelineUnnecessary to processContentInOwnPipeline and
inverts the logic.  An assertion is made that we're in link context when
determining if we need a subpipeline because otherwise the logic was
busted if an XMLTagTk was encountered before a transclusion meta.

Change-Id: I7a04dd1041a6f575bb216ee67c121efe054b55e1