Fix XMLSerializer alias
The "classmap-authoritative" feature doesn't automatically pick up
class aliases, but we can fake it out.

This bug has already been worked around in production, but this
should help document the "correct" way to create an alias if we
need to do this again in the future.

This is a cherry-pick from
I31be0b8cd3b999b397ac8af9686784de3b2b519e, with some tweaks to remove
rich-data specific node dumps.

Bug: T393983
Change-Id: I31be0b8cd3b999b397ac8af9686784de3b2b519e