Replace unnecessary uses of `empty()`
A MediaWiki code convention, enforced via phan, is to avoid the use of
`empty()` unless specifically testing for the presence of a dynamic
propery or array element which might be missing.

Fixing these allowed re-enabling the phan category.  Made a few
types more precise in the process as well, and fixed a type error
this revealed in WikitextSerializer where we were using an empty
array in place of a ParamInfo.

Change-Id: Ica5fe952f3fef12589c0c978c64597b30c1b25f3