Simplify list frame handling in ListHandler
Previously, list frames were pushed/popped from the stack when entering
and exiting a table in a list, and an active list frame was stored in
the $currListFrame var.  That bookkeeping was a little cumbersome.

List frames are only created when there's no active frame and a list is
encountered.  So, entering a table in a list would push the active list
frame on the stack and null out the $currListFrame so a new one could be
created if encountered.  Similarly, when exiting a table in a list, any
active list frame would need to be closed and then a frame popped from
the stack and set as current.

The new logic keeps a boolean of whether we $haveActiveListFrame.  List
frames are pushed onto the stack when list frames are created and popped
when they're closed.  Entering a table in a list toggles
$haveActiveListFrame so that a new list can be created and pushed on the
stack.  Exiting the table in a list will set $haveActiveListFrame so the
top of the stack list frame is active again.  The active list frame is
always at the top of the stack.

Change-Id: If449df7abdd79786087e1fecf24d52ed2c864d58