Update normalize php output method
The mw-headline spans are no longer emitted in the legacy output so the
regexp can't match.  A bunch of tests with only html sections move off
the known failures list.

This fix is an alternate proposal to
I202d892da7d0d1cd90275bbed7504cc4ee3a668c which removed the method
entirely.

Bug: T395234
Change-Id: Iea270754efcfef11aaa8d5f4a2150e00e359a731