Tweak attribute comparison in DomDiff
Ensure that the attribute comparison method is only called on non-null
values, which /should/ be the case naturally, since "missing" attributes
are omitted from the keys array.  However, note that calling ::getDataMw()
and ::getDataParsoid() will create empty objects on each node if they didn't
already exist, which is needless allocation.  We should probably revisit
this in the future to be more careful.  We should also add support here
for other rich attributes, like data-mw-i18n, which are currently being
compared based on their string values.

Change-Id: Ib0033c2253043c4d1b288324e935744853f22427