Create language converter only once for AddRedLinks
To get all language variants of a link, AddRedLinks repeatedly creates a
LanguageConverter for the language of the page. This is inefficient and
it happens to leak, leading to OOMs when the amount of redlinks on a
page is large. This patch extracts the LanguageConverter loading so that
it happens once per AddRedLinks instead.

Bug: T395630
Change-Id: I87e4bf2edadfc0cc838d772cdfc7c64a3e2ee454