Mark or fix unused foreach variables
Analysis tools use the convention that a leading underscore indicates
that the value is expected to be unused.  In some cases the unused
variable can be removed, but in others it serves as useful
documentation of the expected values included in a key/value pair
so I've left the unused variable but added the leading underscore.

Change-Id: Ifc1ad42445e0bd93b022d285c7090ffd135c1eee