Only deduplicate about and annotation ID beginning at a start tag
When we clone a fragment already inside an transclusion or annotation
region, for example when doing cloning from the Active Formatting Element
list during UnpackDOMFragments, we don't want to renumber the about IDs.

But if we have a standalone fragment, with start/end tags for the
transclusion or annotation, then it is safe to renumber the about region.

This was a regression found during round-trip testing, when lint errors
were being generated without DSR information because the about IDs had
been renumbered and they no longer matched their containing transclusion.

Test case added to LinterTest which ought to prevent this sort of
regression in the future.

Follows-Up: I8b824f23024dcbccb3a4acfe1aca4b59cf8589c4
Change-Id: Icacd1bace7de27c2aea16eee90d4de5acf883776