Remove DOMFragment contents when it is cloned for use as a wrapper
We try to reuse the outermost tag on a DOMFragment when we create
a wrapper for it, mostly so that block/inline characteristics are
preserved.  When we create a DOMFragment wrapping another
DOMFragment, this was inadvertently copying the inner fragment's
HTML and triggering an assertion when that HTML was overwritten.

Change-Id: Id4a1fa3a03215cfa219ad06f4cd8938f7e2a2637