Grammar.pegphp: more table parsing rule optimizations
Reduce the number of backtracking entries from 20 to 13 for each pair
of `!!` characters:

1. Use 'newline' instead of 'newlineToken' when we don't actually need
a token created.

2. Use 'optional_spaces' instead of 'optionalSpaceToken' when we don't
actually need a token created.

3. Test inline_breaks earlier in nested_block_in_table

4. Move the `$` to the caller of optional_spaces, instead of the rule
itself, which should enable wikipeg inlining of the rule in the future.

5. Add a test to the start of the `sol` rule to prevent a lot of
unnecessary backtracking cache entries in the common case where we are
not at the start of the line.

Bug: T353874
Change-Id: If363a9ffabd00ff80dc25ced69756e00fda75c52