Regenerate tokenizer with wikipeg 6.0-prerelease
This is regenerated using the latest prerelease for wikipeg 6.0,
git hash 2e1e8db6dd552691502b4d35607a2d027e774d67 in the
wikimedia/wikipeg repo.

This contains optimizations for position tracking, repeated
character classes, and inlining simple rules.  It also contains
bugfixes for backtracking in the presence of reference variables
and register management of sequence expressions.

This version of wikipeg has not changed the PHP runtime, and
so we've bumped only the node dependency in package.json which
is used for generating the tokenizer.  The composer package
version is used only for the PHP runtime and is unmodified.

Change-Id: I107dd5e337cd8d487d73a0e04df33283ec959baf