Grammar.pegphp: Remove some infinite loops
We can't quite enable the 'no infinite loop' checker in wikipeg,
but this brings us closer.

Infinite productions are things like `""*` -- a production which
matches the empty string, repeated any number of times.  In practice
for all the infinite loops currently flagged we are breaking the loop
via other mechanisms, but it's still worth checking that we don't add
any infinite loops during maintenance.

Change-Id: I461875881deccee23830e2c3fa72c48880abb53c