Don't hoist categories outside the <body> in the presence of bare <li>
The LiFixups code was assuming that every list item (li/dd/dt)
was properly nested in an appropriate list container (ul/ol/dl).
However, in the presence of wikitext which has bare <li> items
without a corresponding container, the fixup code could end up
hoisting the category outside of whatever element happened to be
the parent of the <li> -- including the <body>.

When DOM nodes end up outside the body, strange things start to
happen.  The DOMDataUtils load/store code didn't touch the node,
since it was outside of the body, until the ExtractBody stage in
the OutputTransform magically dropped it inside again, but with a
random data-object-id, ultimately leading to a crash because we
found two nodes with the same data-object-id, which should never
happen.

Bug: T390629
Change-Id: Ief1d70b289953f00c382b87c997e8a1f236bece4