Grammar.pegphp: use @ (pick) operator
This depends on I325aae70b93ba550a37b406a252372c86292fbd5 in wikipeg.

Documentation on the new pick operator is at
https://doc.wikimedia.org/wikipeg/master/index.html#autotoc_md24

To summarize:

    *expression1* *expression2* ... *expressionn*

Matches a sequence of expressions and return their match results in an array. Elements of the sequence can be picked by preceding them with the pick operator (@), and only those elements will be returned in the array. If only one element is picked, it is returned directly (not wrapped in an array).

    @ expression

Picks the specified expression in a sequence to return. See the description of a sequence expression above.

Note that sequences with pick operators can be nested, for example:

    foo = @"a" @("b" @"c" "d") "e"

will return ["a", "c"] if it matches.

Change-Id: I1d0aede29e559b3cdcb307b4e526e516fd90b47c