Grammar.pegphp: th and preproc parameters should be reference parameters
When we reference preproc and th in the inline_breaks rule, we should
mark them as reference parameters.  This seems to have been a bug
introduced at the very introduction of reference parameters, in
ea80d1b6854dc586c3dbe5419055bbb1527bb743, although in JavaScript
we didn't strictly have to mark the type in order to pass reference
parameters correctly.

Change-Id: I930090452beeb84d533cea98d18893af0b6714bf