Grammar.pegphp: mark the most common rules for inlining
Depends on Ibb1cc9d3fc1eacc9b97bbba18e5022e906fde31c in peg.

In the case of `inline_breaks` we just set `[cache=false]`
instead of inlining it, which has roughly the same effect
since this is a nullable rule that doesn't contribute
to character class optimizations.

Reduces the number of cache entries/rule invocations for the
simple urltext `http://example.com` from 127 to 33.

Change-Id: Ib2fafcf1c12e06d76669fd9cc4eb362af122961e