DomPageBundle: clarify by-reference output parameters, always provide Env
In comments to I274ab06cc1ab7792f8c02010032d790aff107d51 it was pointed
out that the by-ref output parameters to DomPageBundle::toDom() and
DomPageBundle::toInlineAttributeHtml() could be confusing if the
caller passed in a non-empty array.  Tweak the signature to ensure the
output parameter is set to a newly-created array to avoid this hazard.

In the process of writing test cases for these newly-added arguments,
discovered that ::fromLoadedDocument() doesn't behave well if you
don't provide it an Env.  The comment indicated that *extension-added*
document fragments would be skipped when constructing the id index,
but in fact *all* embedded document fragments, including those from
image captions, language variant markup, expanded attributes, etc,
would be skipped unless an Env was provided.

It turns out that a MockEnv and a SiteConfig are all that's necessary,
so tweak ::fromLoadedDocument to accept a SiteConfig as well, which we
should be easily able to provide from mediawiki core (the SiteConfig
is available via ServiceWiring).

Follows-Up: I274ab06cc1ab7792f8c02010032d790aff107d51
Change-Id: Ic0be74a5b42f3ec0c9dbf359ddda8c835a618e9e