Ensure <mw:editsection> can round-trip with PHP 8.4
In Ia2638e3e692ce5ec3b805384c388d1a7786b8a31 it was discovered that
'<mw:editsection>' was round-tripping as '<editsection>' when
converted to HTML, which was causing tests to fail on PHP 8.4.

This problem was fixed three ways.

The problem was that `Element.createElementNS(..., 'mw:editsection')`
created a node with `localName` set to `editSection` and `prefix`
set to `mw`.  `namespaceURI` was correctly set to the HTML namespace.
XHtmlSerializer didn't handle nodes with non-null `prefix`, and so
fix #1 was to test prefix here and use it in the serialization
where necessary.

But naive `div.innerHTML="<mw:editsection>foo</mw:editsection>"`
doesn't actually result in `prefix` being non-null; instead the
`localName` of the node is set to `mw:editsection`.  It turns
out that using `document.createElement("mw:editsection")`
correctly creates a node with the desired properties.  But
we'd been using `createElementNS` because in PHP < 8.4,
`createElement` creates a node with a null namespace (not
the HTML namespace) which isn't correct either.  The
`suppressHtmlNamespace` option in Remex was added to allow
this incorrect null namespace, since for PHP < 8.0 the DOM
operations were notably faster when namespace was null.

Fix #2 was to have RemexHtml set `suppressHtmlNamespace` to
true by default for PHP 8.4, since using `createElement`
instead of `createElementNS` doesn't actually cause the
namespace to be (incorrectly) null anymore.  And using
`createElement` ensures that the `localName` (correctly)
has a colon in it when appropriate.  (This is
Iccb556c0bca6f9d911f1166704bf2078c314abd3.)

But in order to not to have to wait for a new Remex
release, fix #3 is to set `suppressHtmlNamespace` to
`true` explicitly in the options Parsoid uses for
remex.

A test case is also added to DOMUtilsTest to hopefully
catch this if it regresses.

Change-Id: I3502320cd3be21fcdba220b08618afd47f3f4422