Use SiteConfig for processing attribute embedded DOM/HTML
Instead of an optional ParsoidExtensionAPI, require a mandatory
SiteConfig for ContentUtils::processAttributeEmbedded(Dom|HTML), since
it will skip extension-embedded attributes if SiteConfig is null.

Similarly, require a SiteConfig for DOMDataUtils::usedIdIndex() and
DomPageBundle::fromLoadedDocument(), along with basically anyone who
calls DOMDataUtils::visitAndStoreDataAttribs, since the idIndex is
important for correctness there.

We should also probably mark the first argument of
TagExtensionHandler::processAttributeEmbedded(Dom|HTML)
and PFragmentHandler::processAttributeEmbeddedDom
as deprecated or "do not use" in some way, since we're
always using a mock for that now.  Not sure the best
way to do that deprecation, probably we'd need to
rename the interface method (sigh).

Change-Id: If8d4c72ec875f117082ee60de57a726ac4084832