Don't expand template-like strings returned from the preprocessor
If we always expand template tokens with parsoid fragment prefix, we can
avoid having to expand all template-like strings returned from the
preprocessor.  A bit of refactoring accomplishes this and distinguishes
the case from fragment handlers which expressly want to allow for
expansion.

The test case hits the recursion depth limit without this change.  It
was isolated from visual diff'ing the following page,
https://blk.wikipedia.org/wiki/%E1%80%90%E1%80%99%E1%80%BA%E1%80%B8%E1%80%95%E1%80%9C%E1%80%B1%E1%80%80%E1%80%BA:Law-stub?useparsoid=1

Bug: T385806
Change-Id: I4e3015e5ecaee80441c8e49762e57d4770e9790d