Revert "Ensure <mw:editsection> can round-trip with PHP 8.4"
This reverts some of commit cf12ce3eea0e669858410841e60d68c68e45b3ff.

This is a temporary measure until a more thorough analysis and fix can
be deployed.

What's known is that, after calling out to the legacy parser for
non-native extension output, Parsoid will parse the returned html and
import it to a fragment of the top level doc.  During that import, math
nodes in the mathml xmlns will go from having an empty string '' set as
prefix 'default'.  This only occurs when the legacy php dom
implementation from prior to 8.4 is used.  Possibly, a
DOMCompat::importNode is needed to account for the issue.  It existed
prior to the patch but hadn't been exposed since the prefix was ignored
during xhtml serialization.

Bug: T400060
Change-Id: I9bc66ed5a604f7659503a851cc5ee53e7f4b3dfd