Switch to named when numbered tpl arg has tag with attr
Two tests are added.  The first captures the htmll2wt fix in this patch.
The second, however, shows that, prior to this patch, the html2html mode
of the first test should have been failing but wasn't and will be fixed
in the follow up in I40ba0ef0167324b5e7cfbfb019afa6d7a3245891.

Bug: T400080
Change-Id: I479e28d3318ec8414b7eff40dd34d4dcfef914ff