Fix breaking tpl args on equals in html tags
Resolves a discrepancy between what ends up in the data-mw vs what
actually gets expanded because we send the entire source of a template
to the preprocessor and are subject the logic there, which breaks on the
equals.

Also, in 05654909026593ae76a93a6c3a36dc7a5ef34f77, we started escaping
when serializing with an equals in a positional argument and without
this fix, arguments that are wrongly considered positional are being
converted to named and breaking semantics.  See the revert in
Ibaccad0c978cc79d3d93682e587b06bf520a228f.

Note the failing test is added because pfs should tokenize differently,
they don't have named arguments.

Bug: T400080
Change-Id: I40ba0ef0167324b5e7cfbfb019afa6d7a3245891