Improve performance of smartQuote feature in XHtmlSerializer
This is a follow-up for what I started in I02b7cc2.

This code is called exceptionally often, hundreds of thousands of
times. The new code is significantly faster, because:

1. The old code was executing substr_count two (!) times for every
attribute. But we know that quotes are exceptionally rare in HTML
attributes. The new code is not only faster because str_contains is
faster, it's also executed only half as often as before.

Numbers for an example test run I did:
Before: 125284 calls to substr_count
After:   62642 calls to str_contains
         + 925 calls to str_contains
          + 98 calls to substr_count
       = 63665 calls

2. The function call overhead in PHP is rather large. I think this is
not worth it in this case here, for a single line of code. We don't
loose much. The names of the constants still explain what's going on.

I think the extra unit test is not really needed. There are plenty of
other tests that cover this.

Change-Id: I43cc67e0f993706beb2cd50fa261b822414cce85