Set attrSrc tmp on the rest of the table tags
attrSrc was added to do table fixups in the presence of missing source
range information.  However, it was only added to td/th.  Setting it on
the rest of the table tags allows us to drop the best effort
reconstruction of the source from tag attributes in token stream
patching.

Bug: T400633
Follows-Up: Ifb4d58cce595115b7948a5768f8e48086f8c4039
Change-Id: I3dd47d8a8d74284d9fd4558587215ecceac77245