Shift tokens after parsing to expanded
Shifting after is more robust because some handlers in stage 2 expect
tsr info (TemplateHandler) and so clearing it can lead to crashes.
However, because TemplateHandler itself clears tsr info, subsequent
handlers should be already prepared for missing info.

Some changes in this patch:

* An option to processContentInPipeline to pass in the startRule.
  Previously, a PegTokenizer was constructed and tokenizeAs was called
  directly, rather than doing the tokenizing and token transfroms all in
  one pipeline.

* ParserPipeline used to only create a new frame if we weren't at the
  top level.  Now, if srcText is provided, a new frame is constructed
  regardless.

Meandering:

Rather than setting a srcText, since the source is supposed to be coming
from the current frame in this case, we could pass in srcOffsets to the
pipeline, but there's two issues with that.

One is that the grammar only shifts the tsr of tokens in emitChunk, but
start rules other than start/start_async don't call that, so tokens
won't get shifted in those cases.  We should consider moving the shift
out of the grammar unless the idea of using different start rules in a
parser pipeline isn't sound.

Second, in some cases of the TokenStreamPatcher, we might be clearing
tsr info.  The point of this patch is that we don't want to clear until
after the pipeline is done processing to expanded.

Bug: T401454
Change-Id: Ia9ff0387c2970ea19639984fb9387839a70aba62