Turn on hints to improve the serialization of I18nInfo
This corresponds to core patch
If1fd33408588f16feaaaa5cffde934a22400f8d6 which turns on similar hints
in core.

To break a circular dependency with parser test output in
the Cite extension, which has data-mw-i18n attributes which
would be changed by this patch, we temporarily disable some
tests in the Depends-On patch.  Once this is merged and tagged
into vendor, the cite tests can be updated and those Cite
tests can be re-eanbeld.

Depends-On: If1fd33408588f16feaaaa5cffde934a22400f8d6
Depends-On: I963ee78a5cf10ac37658fceb632d22e3c24652cc
Change-Id: Ic306d26194daf03bcc65b5e5a1e18cb67aa1c046