Skip duplicate JsonSchema validations for performance reasons
This code was doing the same JsonSchema validation for the same
.json file over and over again for every single test in an extension.
This adds up. I can see the difference in an example parser test run
I did. With this optimization it does down from about 1.4s to 1.1s.

Not relevant for production, luckily, but still worth it in my
opinion.

Change-Id: If0f2cd3924476b84efb99068aad93a703e0ea761