Use bracketPairStrippedStr when doing nowiki checks
We shouldn't need to escape these chars when they are in paired brackets
in template args.

The replacement for bracketPairStrippedStr was a bit odd in that only
inner text from the first capture group (square brackets) was retained.
This was most likely an oversight in 082d0c9 but the effect was that
template and language conversion markup were entirely omitted from the
nowiki check.  This patch argues that we can extend that to square
brackets as well and use the stripped string for all the checks.

For example, {{1x|[[Testing=123|t=123]]}}, the equals and bars in the
link don't need escaping.

Change-Id: I83e8a6958d388df54224341acd66e8e2490a90d5