Revert "Reapply "Switch to named when numbered tpl arg has tag with attr""
This reverts commit 05654909026593ae76a93a6c3a36dc7a5ef34f77.

Reason for revert: RT testing regression: `{{tpl|..<div class="center">..</div> ..}}` being serialized as `{{tpl|1=..<div class="center">..</div> ..}}`  because of the presence of the '=' can change output because previously `{{tpl}}` which might have processed a `{{{1|}}}` arg as empty will process it as a non-empty string. As it happens, in this case, what the kowiki editors wanted is probably the Parsoid-serialized version in the first place (looking at other uses of the template on the page), but strictly as far as wt2wt goes, this is a dirty semantic diff.

I40ba0ef0167324b5e7cfbfb019afa6d7a3245891 will likely fix it once it is mergeable.

Change-Id: Ibaccad0c978cc79d3d93682e587b06bf520a228f