Rename DOMCompat::nodeName() to DOMUtils::nodeName()
Our implementation doesn't actually return a standards-compliant result,
which would be uppercase.  Instead we return what PHP 7.x returns, which
is lowercase.  In the future when we migrate to PHP 8.4+ the native
return value will be uppercase and we will probably want to test against
that instead.

So rename DOMCompat::nodeName() to DOMUtils::nodeName() so we can
deprecate DOMCompat::nodeName() and eventually bring it back in the
future with an *actually* standards-compliant return value (aka
uppercase).

Change-Id: I34b9f09972439ae275b10b7d1914d36028cd670e