Consider all pf arguments as positional
Parser functions in the legacy parser do not support named parameters,
see T204307.  Representing them as named in the data-mw is inconsistent
with how they are used when rendering and can lead to deduplication
resulting in semantic differences, for each,
{{#tag:pre|class=123|class=123}} roundtrips as {{#tag:pre|class=123}},
which drops the tag class attribute, only keeping the string content.

However, v3 parser functions in development will support named
parameters, as described in
[[mw:Parsoid/MediaWiki_DOM_spec/Parser_Functions]], and larger changes
to parser function representation will be coming in the next major
version of the spec.

Note that code out there should already be able to handle positional
arguments, they'll just be getting more of them.  The current spec will
the updated to reflect this change.

Bug: T400080
Bug: T204307
Depends-On: I3ea6d3a0c117a9e477c27b390700044a016bb6a6
Depends-On: Ifa07954d0c7cb67945d9b8a2f82453852aa903a6
Depends-On: Id51b349c237e84eb129986cb8b480790ee251a27
Change-Id: I4e0a93cb2ff52f71d927ed9ef98c2e1305de55a8