Eliminate bad dsr for br logs from rt testing
The following edit fixed the issue for the set of tests we were running
the regression script on,
https://fr.wikipedia.org/w/index.php?title=Icare&diff=228920011&oldid=228896434

What was happening was the unclosed quotes we causing AFE to leave
around some empty formatting nodes that, after normalization, resulted
in the insertion of a br.  In the normal usage of the normalizer, the br
is marked as inserted so dsr info isn't expected on the node.  However,
the roundtrip testing script assumes dsr for all nodes that aren't
encapsulated, thus the normalizer adding a node w/o dsr info is
unexpected.

Change-Id: I41729b192806d5db88c52b4b0406b12c37f63cc9