Use new 'template3' token for `{{#...}}` constructs
This patch maintains legacy DOM output for "old" parser functions,
using the serialization compatiblity code introduced in
I1c78a6d67e63133bdd338d64e62ca2b19374762a, but parses `{{#...}}`
constructs with a new "template3" token which represents
arguments as list<string|PreprocTk>; that is, as
preprocessed-but-otherwise-untokenized text.

Native PFragmentHandlers are always tokenized with `template3` and
the "native" (standalone) parserfunctions and templates are always
tokenized with `template`; we retokenize the source string if necessary
to guarantee this, although this won't happen in production since
all native PFragmentHandlers use the `{{#...}}` form.  (We don't
retokenize other `{{#...}}` constructs in integrated mode; we instead
ask mediawiki-core to expand them directly from the source of the
`template3` token.)

With the new representation of parser function arguments,
PFragmentHandlers gain support for named arguments (T390344)
and experimental support for varargs (T204366#10772988), although
varargs are currently blocked by the lack of core preprocessor
support for invoking PFragmentHandlers (T390345).

We expect most uses of the `template` token can eventually be
migrated to the `template3` token (T394836).

Bug: T394834
Bug: T390344
Change-Id: Ibf3fc177713ec487bbace32a4b14b5a103358b5e